+++ wulu/src/commands/money.js
@@ -49,7 +49,7 @@ function money() {
 
     takemoney(target, room, user) {
       if (!user.can('takemoney')) return;
-      if (!target || target.indexOf(',') < 0) return this.parse('/help takemoney');
+      if (!target || target.indexOf(',') < 0) return;
 
       let parts = target.split(',');
       this.splitTarget(parts[0]);
@@ -71,7 +71,7 @@ function money() {
 
     transfer: 'transfermoney',
     transfermoney(target, room, user) {
-      if (!target || target.indexOf(',') < 0) return this.parse('/help takemoney');
+      if (!target || target.indexOf(',') < 0) return this.sendReply('/transfer [user], [amount] - Transfer a certain amount of money to a user.');
 
       let parts = target.split(',');
       this.splitTarget(parts[0]);